# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  auth.DeleteAccountInput:
    properties:
      confirm:
        type: string
      password:
        type: string
    type: object
...
  do.AnnouncementStatus:
    enum:
    - 0
    - 1
    - 2
    type: integer
    x-enum-comments:
      AnnouncementStatusArchived: 已归档
      AnnouncementStatusDraft: 草稿
      AnnouncementStatusPublished: 已发布
    x-enum-descriptions:
    - 草稿
    - 已发布
    - 已归档
    x-enum-varnames:
    - AnnouncementStatusDraft
    - AnnouncementStatusPublished
    - AnnouncementStatusArchived
  do.AnnouncementType:
    enum:
    - 0
    - 1
    - 2
    - 3
    type: integer
    x-enum-comments:
      AnnouncementTypeEmergency: 紧急公告
      AnnouncementTypeEvent: 活动公告
      AnnouncementTypeImportant: 重要公告
      AnnouncementTypeNormal: 普通公告
    x-enum-descriptions:
    - 普通公告
    - 重要公告
    - 紧急公告
    - 活动公告
    x-enum-varnames:
    - AnnouncementTypeNormal
    - AnnouncementTypeImportant
    - AnnouncementTypeEmergency
    - AnnouncementTypeEvent
...
  request.CreateAnnouncement:
    properties:
      board_id:
        type: integer
      content:
        type: string
      cover:
        type: string
      expired_at:
        type: string
      is_global:
        type: boolean
      is_pinned:
        type: boolean
      published_at:
        type: string
      status:
        $ref: '#/definitions/do.AnnouncementStatus'
      summary:
        type: string
      title:
        maxLength: 200
        minLength: 1
        type: string
      type:
        allOf:
        - $ref: '#/definitions/do.AnnouncementType'
        enum:
        - 0
        - 1
        - 2
        - 3
    required:
    - content
    - title
    - type
    type: object
  request.ListAnnouncements:
    properties:
      board_id:
        type: integer
      end_time:
        type: string
      is_global:
        type: boolean
      is_pinned:
        type: boolean
      keyword:
        type: string
      page:
        minimum: 1
        type: integer
      page_size:
        maximum: 100
        minimum: 1
        type: integer
      start_time:
        type: string
      status:
        allOf:
        - $ref: '#/definitions/do.AnnouncementStatus'
        enum:
        - 0
        - 1
        - 2
      type:
        allOf:
        - $ref: '#/definitions/do.AnnouncementType'
        enum:
        - 0
        - 1
        - 2
        - 3
    type: object
...
  request.UpdateAnnouncement:
    properties:
      board_id:
        type: integer
      content:
        type: string
      cover:
        type: string
      expired_at:
        type: string
      is_global:
        type: boolean
      is_pinned:
        type: boolean
      published_at:
        type: string
      status:
        allOf:
        - $ref: '#/definitions/do.AnnouncementStatus'
        enum:
        - 0
        - 1
        - 2
      summary:
        type: string
      title:
        type: string
      type:
        allOf:
        - $ref: '#/definitions/do.AnnouncementType'
        enum:
        - 0
        - 1
        - 2
        - 3
    type: object
...
  user.LoginInput:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  user.RegisterInput:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
      username:
        maxLength: 50
        minLength: 2
        type: string
    required:
    - email
    - password
    - username
    type: object
  vo.BasicResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      request_id:
        type: string
      timestamp:
        type: integer
      trace_id:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: 一个基于 Gin 的论坛系统 API
  title: Tiny Forum API
  version: "1.0"
paths:
  /admin/announcements:
    post:
      consumes:
      - application/json
      description: 管理员创建一条新公告，需要认证及管理员权限
      parameters:
      - description: 公告信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.CreateAnnouncement'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            $ref: '#/definitions/vo.BasicResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/vo.BasicResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/vo.BasicResponse'
        "403":
          description: 无权限（非管理员）
          schema:
            $ref: '#/definitions/vo.BasicResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/vo.BasicResponse'
      security:
      - ApiKeyAuth: []
      summary: 创建公告
      tags:
      - 公告管理
  /admin/announcements/{id}:
    delete:
      consumes:
      - application/json
      description: 管理员根据ID删除公告，需要认证及管理员权限
      parameters:
      - description: 公告ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/vo.BasicResponse'
        "400":
          description: 参数错误（无效的公告ID）
          schema:
            $ref: '#/definitions/vo.BasicResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/vo.BasicResponse'
        "403":
          description: 无权限（非管理员）
          schema:
            $ref: '#/definitions/vo.BasicResponse'
        "404":
          description: 公告不存在
          schema:
            $ref: '#/definitions/vo.BasicResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/vo.BasicResponse'
      security:
      - ApiKeyAuth: []
      summary: 删除公告
      tags:
      - 公告管理
    put:
      consumes:
      - application/json
      description: 更新公告信息
      parameters:
      - description: 公告ID
        in: path
        name: id
        required: true
        type: integer
      - description: 公告信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.UpdateAnnouncement'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/vo.BasicResponse'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/vo.BasicResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/vo.BasicResponse'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/vo.BasicResponse'
        "404":
          description: 公告不存在
          schema:
            $ref: '#/definitions/vo.BasicResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/vo.BasicResponse'
      security:
      - ApiKeyAuth: []
      summary: 更新公告
      tags:
      - 公告管理
  /admin/announcements/{id}/archive:
    post:
      consumes:
      - application/json
      description: 将公告状态设为已归档
      parameters:
      - description: 公告ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 归档成功
          schema:
            $ref: '#/definitions/vo.BasicResponse'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/vo.BasicResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/vo.BasicResponse'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/vo.BasicResponse'
        "404":
          description: 公告不存在
          schema:
            $ref: '#/definitions/vo.BasicResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/vo.BasicResponse'
      security:
      - ApiKeyAuth: []
      summary: 归档公告
      tags:
      - 公告管理
  /admin/announcements/{id}/pin:
    put:
      consumes:
      - application/json
      description: 设置公告的置顶状态
      parameters:
      - description: 公告ID
        in: path
        name: id
        required: true
        type: integer
      - description: 置顶状态
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 操作成功
          schema:
            $ref: '#/definitions/vo.BasicResponse'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/vo.BasicResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/vo.BasicResponse'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/vo.BasicResponse'
        "404":
          description: 公告不存在
          schema:
            $ref: '#/definitions/vo.BasicResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/vo.BasicResponse'
      security:
      - ApiKeyAuth: []
      summary: 置顶/取消置顶公告
      tags:
      - 公告管理
  /admin/announcements/{id}/publish:
    post:
      consumes:
      - application/json
      description: 将公告状态设为已发布
      parameters:
      - description: 公告ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 发布成功
          schema:
            $ref: '#/definitions/vo.BasicResponse'
        "400":
          description: 参数错误或发布时间无效
          schema:
            $ref: '#/definitions/vo.BasicResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/vo.BasicResponse'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/vo.BasicResponse'
        "404":
          description: 公告不存在
          schema:
            $ref: '#/definitions/vo.BasicResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/vo.BasicResponse'
      security:
      - ApiKeyAuth: []
      summary: 发布公告
      tags:
      - 公告管理
  /admin/announcements/list:
    get:
      consumes:
      - application/json
      description: 管理员列出所有公告，需要认证及管理员权限
      parameters:
      - description: 公告信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.ListAnnouncements'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            $ref: '#/definitions/vo.BasicResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/vo.BasicResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/vo.BasicResponse'
        "403":
          description: 无权限（非管理员）
          schema:
            $ref: '#/definitions/vo.BasicResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/vo.BasicResponse'
      security:
      - ApiKeyAuth: []
      summary: 列出公告
      tags:
      - 公告管理
...
  /auth/account:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 注销请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.DeleteAccountInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/vo.BasicResponse'
      security:
      - ApiKeyAuth: []
      summary: 用户注销账户（软删除）
      tags:
      - 验证管理
  /auth/account/deletion:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/vo.BasicResponse'
      security:
      - ApiKeyAuth: []
      summary: 获取注销状态
      tags:
      - 验证管理
  /auth/account/permanent:
    delete:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/vo.BasicResponse'
      security:
      - ApiKeyAuth: []
      summary: 确认永久删除账户
      tags:
      - 验证管理
  /auth/account/restore:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/vo.BasicResponse'
      security:
      - ApiKeyAuth: []
      summary: 取消注销账户
      tags:
      - 验证管理
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: 登录信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/user.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/vo.BasicResponse'
      summary: 用户登录
      tags:
      - 验证管理
  /auth/logout:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/vo.BasicResponse'
      security:
      - ApiKeyAuth: []
      summary: 用户登出
      tags:
      - 验证管理
...
  /auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: 注册信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/user.RegisterInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/vo.BasicResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/vo.BasicResponse'
      summary: 用户注册
      tags:
      - 验证管理
...
securityDefinitions:
  ApiKeyAuth:
    description: Type "Bearer" followed by a space and the JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"