# backend/internal/handler/announcement/update.go
package announcement

import (
	"tiny-forum/internal/model/request"
	"tiny-forum/pkg/response"

	"github.com/gin-gonic/gin"
)

// @Summary 更新公告
// @Description 更新公告信息
// @Tags 公告管理
// @Accept json
// @Produce json
// @Security ApiKeyAuth
// @Param id path int true "公告ID"
// @Param body body request.UpdateAnnouncement true "公告信息"
// @Success 200 {object} vo.BasicResponse"更新成功"
// @Failure 400 {object} vo.BasicResponse"参数错误"
// @Failure 401 {object} vo.BasicResponse"未授权"
// @Failure 403 {object} vo.BasicResponse"无权限"
// @Failure 404 {object} vo.BasicResponse"公告不存在"
// @Failure 500 {object} vo.BasicResponse"服务器内部错误"
// @Router /admin/announcements/{id} [put]
//
// Deprecated: 迁移到 adminHandler.UpdateAnnouncements
func (h *AnnouncementHandler) Update(c *gin.Context) {
	id, ok := parseAnnouncementID(c)
	if !ok {
		return
	}
	var req request.UpdateAnnouncement
	if err := c.ShouldBindJSON(&req); err != nil {
		response.BadRequest(c, err.Error())
		return
	}
	userID := c.GetUint("user_id")
	if err := h.service.Update(c.Request.Context(), id, &req, userID); err != nil {
		response.HandleError(c, err)
		return
	}
	response.Success(c, nil)
}

// @Summary 发布公告
// @Description 将公告状态设为已发布
// @Tags 公告管理
// @Accept json
// @Produce json
// @Security ApiKeyAuth
// @Param id path int true "公告ID"
// @Success 200 {object} vo.BasicResponse"发布成功"
// @Failure 400 {object} vo.BasicResponse"参数错误或发布时间无效"
// @Failure 401 {object} vo.BasicResponse"未授权"
// @Failure 403 {object} vo.BasicResponse"无权限"
// @Failure 404 {object} vo.BasicResponse"公告不存在"
// @Failure 500 {object} vo.BasicResponse"服务器内部错误"
// @Router /admin/announcements/{id}/publish [post]
//
// Deprecated: 迁移到 adminHandler.PublishAnnouncements
func (h *AnnouncementHandler) Publish(c *gin.Context) {
	id, ok := parseAnnouncementID(c)
	if !ok {
		return
	}
	userID := c.GetUint("user_id")
	if err := h.service.Publish(c.Request.Context(), id, userID); err != nil {
		response.HandleError(c, err)
		return
	}
	response.Success(c, nil)
}

// @Summary 归档公告
// @Description 将公告状态设为已归档
// @Tags 公告管理
// @Accept json
// @Produce json
// @Security ApiKeyAuth
// @Param id path int true "公告ID"
// @Success 200 {object} vo.BasicResponse"归档成功"
// @Failure 400 {object} vo.BasicResponse"参数错误"
// @Failure 401 {object} vo.BasicResponse"未授权"
// @Failure 403 {object} vo.BasicResponse"无权限"
// @Failure 404 {object} vo.BasicResponse"公告不存在"
// @Failure 500 {object} vo.BasicResponse"服务器内部错误"
// @Router /admin/announcements/{id}/archive [post]
//
// Deprecated: 迁移到 adminHandler.ArchiveAnnouncements
func (h *AnnouncementHandler) Archive(c *gin.Context) {
	id, ok := parseAnnouncementID(c)
	if !ok {
		return
	}
	userID := c.GetUint("user_id")
	if err := h.service.Archive(c.Request.Context(), id, userID); err != nil {
		response.HandleError(c, err)
		return
	}
	response.Success(c, nil)
}

// @Summary 置顶/取消置顶公告
// @Description 设置公告的置顶状态
// @Tags 公告管理
// @Accept json
// @Produce json
// @Security ApiKeyAuth
// @Param id path int true "公告ID"
// @Param body body object true "置顶状态" example({"pinned": true})
// @Success 200 {object} vo.BasicResponse"操作成功"
// @Failure 400 {object} vo.BasicResponse"参数错误"
// @Failure 401 {object} vo.BasicResponse"未授权"
// @Failure 403 {object} vo.BasicResponse"无权限"
// @Failure 404 {object} vo.BasicResponse"公告不存在"
// @Failure 500 {object} vo.BasicResponse"服务器内部错误"
// @Router /admin/announcements/{id}/pin [put]
//
// Deprecated: 迁移到 adminHandler.PinAnnouncements
func (h *AnnouncementHandler) Pin(c *gin.Context) {
	id, ok := parseAnnouncementID(c)
	if !ok {
		return
	}
	var req struct {
		Pinned bool `json:"pinned"`
	}
	if err := c.ShouldBindJSON(&req); err != nil {
		response.BadRequest(c, err.Error())
		return
	}
	userID := c.GetUint("user_id")
	if err := h.service.Pin(c.Request.Context(), id, req.Pinned, userID); err != nil {
		response.HandleError(c, err)
		return
	}
	response.Success(c, nil)
}